refactor: extract the Tauri-free DMX core into lux-engine
New crates/lux-engine holds what the desktop app and the headless node
share: the DmxSink trait and the sACN/E1.31 sender (now with per-sender
priority and source name), universe overlay semantics as a plain struct,
the remote-control route/gate decisions, the webpki PEM bundle, and the
id-token sub reader. The desktop consumes them through re-exports with
behavior unchanged (its sACN sender still announces as lux at the
default priority 100); the moved unit tests travel to the crate.
