+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 members = [
     "apps/desktop/src-tauri",
     "crates/lux-auth",
+    "crates/lux-engine",
     "crates/lux-wire",
     "services/bot",
     "services/iot-authorizer",
